pushing streamlit updated script and dockerfile
pushing streamlit updated script and dockerfile for streamlit
