+++ Streamlit/dockerfile
@@ -1,3 +1,26 @@
+# # Use an official Python runtime as a parent image
+# FROM python:3.8-slim
+
+# # Set the working directory in the container
+# WORKDIR /app
+
+# # Copy the current directory contents into the container at /app
+# COPY . /app
+
+# # Install any needed packages specified in requirements.txt
+# RUN pip install --no-cache-dir -r requirements.txt
+
+# # Make port 8501 available to the world outside this container
+# EXPOSE 8501
+
+# # Define environment variable for Streamlit
+# ENV LC_ALL=C.UTF-8
+# ENV LANG=C.UTF-8
+
+# # Run app.py when the container launches
+# CMD ["streamlit", "run", "--server.port", "8501", "app.py"]
+
+
 FROM python:3.10.12
 
 WORKDIR /Streamlit
@@ -6,6 +29,7 @@ COPY requirements.txt .
 RUN pip install -r requirements.txt
 
 COPY app.py .
+COPY plugins /Streamlit/plugins 
 
 EXPOSE 8503
